fix: check if blocks registry still empty before destroy ctx

+++ src/lit/LitBlock.ts
@@ -190,6 +190,9 @@ export class LitBlock extends LitBlockBase {
 
     if (blocksRegistry?.size === 0) {
       setTimeout(() => {
+        if (!this.isConnected || blocksRegistry?.size > 0) {
+          return;
+        }
         // Destroy global context after all blocks are destroyed and all callbacks are run
         this.destroyCtxCallback();
       }, 0);
